feat: remove markdown parsers
