+++ src/editors/data/services/cms/api.js
@@ -118,8 +118,7 @@ export const apiMethods = {
         couseKey: learningContextId,
         has_changes: true,
         id: blockId,
-        // metadata: { display_name: title, markdown: content.markdown },
-        metadata: { display_name: title, ...content },
+        metadata: { display_name: title, ...content.settings },
       };
     } else if (blockType === 'video') {
       const {
